Isolating dbus-rs usage while fixing connection functionality
A new module is introduced with the intention of creating additional facade isolating the usage of dbus-rs. All returned values and arguments of the new module API are standard Rust structs or our own structs. Internally usage of dbus-rs is improved by using different language features. More complete error handling is introduced with the new module.