All the code using now the `manager` d-bus facade
We are still using the old D-Bus handling code together with the new one and now we are cleaning it all up and making it consistent. I also renamed the `general` module to `status` module, since at the end of the cleanup this is the functionality which ended up there.